Added utils function, tidied code
