+++ src/get_averages.py
@@ -1,46 +1,19 @@
-import sys
 import numpy as np
-import time
-import ase
-from ase import io
-from ase.io import read
-
-import basis
-sys.path.insert(0, './')
 import inp
+from utils import read_system
 
-# read species
-species = inp.species
-
-# read basis
-[lmax,nmax] = basis.basiset(inp.dfbasis)
-
-
-# read system
-xyzfile = read(inp.filename,":")
-ndata = len(xyzfile)
-
-#======================= system parameters
-atomic_symbols = []
-atomic_valence = []
-natoms = np.zeros(ndata,int) 
-for i in range(len(xyzfile)):
-    atomic_symbols.append(xyzfile[i].get_chemical_symbols())
-    atomic_valence.append(xyzfile[i].get_atomic_numbers())
-    natoms[i] = int(len(atomic_symbols[i]))
-natmax = max(natoms)
-
+spelist, lmax, nmax, llmax, nnmax, ndata, atomic_symbols, natoms, natmax = read_system()
 
 avcoefs = {}
 nat_per_species = {}
-for spe in species:
+for spe in spelist:
     nat_per_species[spe] = 0
     avcoefs[spe] = np.zeros(nmax[(spe,0)],float)
 
 print("computing averages...")
 for iconf in range(ndata):
     atoms = atomic_symbols[iconf]
-    coefs = np.load(inp.path2qm+"coefficients/coefficients_conf"+str(iconf)+".npy")
+    coefs = np.load(inp.path2qm+inp.coefdir+"coefficients_conf"+str(iconf)+".npy")
     i = 0
     for iat in range(natoms[iconf]):
         spe = atoms[iat] 
@@ -52,6 +25,6 @@ for iconf in range(ndata):
                        avcoefs[spe][n] += coefs[i]
                     i += 1
 
-for spe in species:
+for spe in spelist:
     avcoefs[spe] /= nat_per_species[spe]
     np.save("averages_"+str(spe)+".npy",avcoefs[spe])
